Removed the class PiBX_AST_Value since it was just a duplication of PiBX_AST_TypeAttribute.
When two classes stand for the same semantical value, one of them cannot justify its existence.
It's just a matter of perspective. When you create an AST from a schema, you speak of a TypeAttribute.
When you create an AST from a binding file, you speak of a "value". But this value is exactly a TypeAttribute - just in another context.